Remove RunHours.IsNullOrWhiteSpace() from Stop()

+++ RoboSharp/RoboCommand.cs
@@ -435,13 +435,15 @@ namespace RoboSharp
         /// <summary>Kill the process</summary>
         public void Stop()
         {
-            if (tokenSource != null)
+            //Note: This previously checked for CopyOptions.RunHours.IsNullOrWhiteSpace() == TRUE prior to issuing the stop command
+	    //If the removal of that check broke your application, please create a new issue thread on the repo.
+	    if (tokenSource != null)
             {
                 if (!tokenSource.IsCancellationRequested) tokenSource.Cancel(true);
             }
-            else if (process != null && CopyOptions.RunHours.IsNullOrWhiteSpace() && !hasExited)
+            else if (process != null)
             {
-                process.Kill();
+                if (!hasExited) process.Kill();
                 hasExited = true;
                 process.Dispose();
                 process = null;
